Conflicts:
	upload/admin/controller/common/menu.php
	upload/admin/language/english/common/menu.php

+++ upload/admin/controller/payment/bluepay_redirect.php
@@ -311,7 +311,7 @@ class ControllerPaymentBluepayredirect extends Controller {
 				$json['data']['created'] = date("Y-m-d H:i:s");
 				$json['data']['amount'] = $this->request->post['amount'];
 				$json['data']['release_status'] = $release_status;
-				$json['data']['total'] = (double)$total_released;
+				$json['data']['total'] = (float)$total_released;
 				$json['error'] = false;
 			} else {
 				$json['error'] = true;
@@ -366,8 +366,8 @@ class ControllerPaymentBluepayredirect extends Controller {
 				$json['data'] = array();
 				$json['data']['created'] = date("Y-m-d H:i:s");
 				$json['data']['amount'] = $this->request->post['amount'] * -1;
-				$json['data']['total_released'] = (double)$total_released;
-				$json['data']['total_rebated'] = (double)$total_rebated;
+				$json['data']['total_released'] = (float)$total_released;
+				$json['data']['total_rebated'] = (float)$total_rebated;
 				$json['data']['rebate_status'] = $rebate_status;
 				$json['error'] = false;
 			} else {
